Clean up and make ts ports work. Apparently you cannot nest an enum in a class, so enums have to be global! What a screwed up OO language.
