+++ sql/mysql/Oracle/Antlr4ng/MySQLParserBase.ts
@@ -4,6 +4,7 @@
 
 import { Parser, TokenStream } from "antlr4ng";
 import SqlMode from "./SqlMode.js";
+import SqlModes from "./SqlModes.js";
 
 export abstract class MySQLParserBase extends Parser {
 
@@ -17,7 +18,7 @@ export abstract class MySQLParserBase extends Parser {
     constructor(input: TokenStream) {
         super(input);
         this.serverVersion = 80200;
-        this.sqlModeFromString("ANSI_QUOTES");
+        this.sqlModes = SqlModes.sqlModeFromString("ANSI_QUOTES");
     }
 
     /**
@@ -49,32 +50,5 @@ export abstract class MySQLParserBase extends Parser {
     public isSelectStatementWithInto(): boolean
     {
         return this.serverVersion >= 80024 && this.serverVersion < 80031;
-    }
-
-    /**
-     * Converts a mode string into individual mode flags.
-     *
-     * @param modes The input string to parse.
-     */
-    public sqlModeFromString(modes: string): void {
-        this.sqlModes = new Set<SqlMode>();
-
-        const parts = modes.toUpperCase().split(",");
-        parts.forEach((mode: string) => {
-            if (mode === "ANSI" || mode === "DB2" || mode === "MAXDB" || mode === "MSSQL" || mode === "ORACLE" ||
-                mode === "POSTGRESQL") {
-                this.sqlModes.add(SqlMode.AnsiQuotes).add(SqlMode.PipesAsConcat).add(SqlMode.IgnoreSpace);
-            } else if (mode === "ANSI_QUOTES") {
-                this.sqlModes.add(SqlMode.AnsiQuotes);
-            } else if (mode === "PIPES_AS_CONCAT") {
-                this.sqlModes.add(SqlMode.PipesAsConcat);
-            } else if (mode === "NO_BACKSLASH_ESCAPES") {
-                this.sqlModes.add(SqlMode.NoBackslashEscapes);
-            } else if (mode === "IGNORE_SPACE") {
-                this.sqlModes.add(SqlMode.IgnoreSpace);
-            } else if (mode === "HIGH_NOT_PRECEDENCE" || mode === "MYSQL323" || mode === "MYSQL40") {
-                this.sqlModes.add(SqlMode.HighNotPrecedence);
-            }
-        });
     }
 }
