+++ sql/mysql/Oracle/Antlr4ng/SqlMode.ts
@@ -1,11 +1,11 @@
 /** SQL modes that control parsing behavior. */
 enum SqlMode {
-    NoMode,
-    AnsiQuotes,
-    HighNotPrecedence,
-    PipesAsConcat,
-    IgnoreSpace,
-    NoBackslashEscapes,
+	NoMode,
+	AnsiQuotes,
+	HighNotPrecedence,
+	PipesAsConcat,
+	IgnoreSpace,
+	NoBackslashEscapes
 }
 
 export default SqlMode;
